chatbot-nutricional-inteligente/requirements.txt
streamlit
langchain
langchain-openai
langchain-community
faiss-cpu
python-dotenv
tiktoken
# Flask==2.1.2
# fastapi==0.75.0
# uvicorn==0.17.0
# numpy==1.21.6
# pandas==1.3.5
# scikit-learn==1.0.2
# torch>=1.10.2
# transformers
# sqlalchemy==1.4.27
# pytest==6.2.5
# pydantic==1.9.0